Nueva interfaz genérica que es llamada desde Operations para convertir a unidades estándares. Renombre a frontend.business.  Se agrega el efecto responsivo en las perillas. Se modifica el plugin del vendor para disminuir por 10px el tamaño de los dígitos. Implementación de Float en los modelos de DTO y en el modelo BusinessView del Frontend.Business. Nuevas clases abstractas para definir medidas en el servicio de backend.

+++ hubble.backend.business.services/src/main/java/hubble/backend/business/services/interfaces/operations/AvailabilityOperations.java
@@ -8,8 +8,8 @@ import java.util.List;
 
 public interface AvailabilityOperations extends AverageOperationsBase<ApplicationAvgDto> {
 
-    public int calculateAverage(List<AvailabilityStorage> availabilityStorageList);
+    public Float calculateAverage(List<AvailabilityStorage> availabilityStorageList);
 
-    public MonitoringFields.STATUS calculateStatus(ApplicationDto appAvg, Integer avgAvailability);
+    public MonitoringFields.STATUS calculateStatus(ApplicationDto appAvg, Float avgAvailability);
 
 }
